feat: Leetcode merge two sorted list
This is the solution to leetcodes merge to sorted list problem.
The pytest will fail on this one.  But it is a working solution.
It is not the optimal solution as it only got:

Runtime: 72 ms, faster than 21.46% of Python3 online submissions for Merge Two Sorted Lists.
Memory Usage: 13.9 MB, less than 32.66% of Python3 online submissions for Merge Two Sorted Lists.

But I wanted to do something different then the other solutions.